# Optimize the cse benchmark