--- a/cmake/checkstdtostring.cpp
+++ b/cmake/checkstdtostring.cpp
@@ -1,0 +1,5 @@
+#include <string>
+#include <iostream>
+int main() {
+   std::cout << std::to_string(0) << std::endl;
+}