--- a/symengine/basic.cpp
+++ b/symengine/basic.cpp
@@ -45,4 +45,9 @@ const char *get_version()
     return SYMENGINE_VERSION;
 }
 
+bool is_a_Atom(const Basic &b)
+{
+    return is_a_Number(b) or is_a<Symbol>(b) or is_a<Constant>(b);
+}
+
 } // SymEngine