Misc optimizations figured out after looking at profiling data for the `large_exprs.TimeLargeExpressionOperations.time_cse` benchmark.

This PR is `2.11 ms` whereas master is `3.76 ms`

This PR also improves other benchmarks.
`expand` benchmark is `768 ms` whereas master is `841 ms`.