@certik 2017-09-30
That's amazing that you managed to speedup the expand benchmark. I thought we reached a peak there already. I'll have a look tomorrow.Sent from my mobile phone.On Sep 29, 2017 9:47 PM, Isuru Fernando <notifications@github.com> wrote:Ping @certik

—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or mute the thread.