@certik 2017-09-30
I ran benchmarks, and I can definitely see an improvement. So I am +1 to merge as it is.